separated toolkit files and added secret reading RBAC role
