+++ config/rbac/role.yaml
@@ -5,6 +5,13 @@ metadata:
   name: manager-role
 rules:
 - apiGroups:
+  - ""
+  resources:
+  - secrets
+  verbs:
+  - get
+  - list
+- apiGroups:
   - tanuu.dev
   resources:
   - nodegroupclaims
